% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\docType{data}
\name{col_type}
\alias{col_type}
\title{Expected Column Types}
\format{
## `col_type`
A data frame with 24 rows and 2 columns:
\describe{
  \item{column}{Column name}
  \item{type}{Data class}
  ...
}
}
\usage{
col_type
}
\description{
A data dictionary for expected column types in the import file
}
\keyword{datasets}
