% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{read_clean_file}
\alias{read_clean_file}
\alias{clean_file}
\title{Read and Clean File}
\usage{
clean_file(file_path = NULL)
}
\arguments{
\item{file_path}{Should be a path on your computer to the SWAN scores}
}
\value{
A clean data frame ready for t-scores
}
\description{
This function reads and cleans the uploaded SWAN scores file
}
