% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{run_longitudinal_model}
\alias{run_longitudinal_model}
\title{Runs the model and creates t-scores}
\usage{
run_longitudinal_model(df = NULL)
}
\arguments{
\item{df}{should be a data.frame from [build_summary()]}
}
\value{
A data frame with t-scores
}
\description{
This model is best used for longitudinal scores. It adjusts for gender, age, and respondent.
This model is suited for looking at changes across time.
Use the dataframe from [build_summary()] to produce t-scores
}
