% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{run_model}
\alias{run_model}
\title{Runs the model and creates t-scores}
\usage{
run_model(df = NULL)
}
\arguments{
\item{df}{should be a data.frame from [build_summary()]}
}
\value{
A data frame with t-scores
}
\description{
Use the dataframe from [build_summary()] to produce t-scores
}
