% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_swan_tscores.R
\name{get_swan_tscores}
\alias{get_swan_tscores}
\title{Run analysis on SWAN raw values to return t-scores}
\usage{
get_swan_tscores(file = NULL, output_folder = NULL)
}
\arguments{
\item{file}{Pathway to formatted raw SWAN scores}

\item{output_folder}{Optional parameter, to export a csv file of the t-scores}
}
\value{
table with t-scores attached to raw swan values
}
\description{
Write more detailed when finished
}
